cube: move `generic.Cube` to `core`
