+++ base/src/main/java/org/aya/core/term/ElimTerm.java
@@ -7,7 +7,6 @@ import kala.collection.mutable.MutableList;
 import kala.collection.mutable.MutableMap;
 import org.aya.core.visitor.Subst;
 import org.aya.generic.Arg;
-import org.aya.generic.Cube;
 import org.aya.util.distill.DistillerOptions;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
@@ -52,7 +51,7 @@ public sealed interface ElimTerm extends Term {
   record PathApp(
     @NotNull Term of,
     @NotNull ImmutableSeq<Arg<@NotNull Term>> args,
-    @NotNull Cube<Term> cube
+    @NotNull FormTerm.Cube cube
   ) implements ElimTerm {
   }
