+++ base/src/main/java/org/aya/tyck/LittleTyper.java
@@ -9,7 +9,6 @@ import org.aya.core.visitor.DeltaExpander;
 import org.aya.core.visitor.Subst;
 import org.aya.generic.Arg;
 import org.aya.generic.Constants;
-import org.aya.generic.Cube;
 import org.aya.generic.util.NormalizeMode;
 import org.aya.guest0x0.cubical.Partial;
 import org.aya.tyck.env.LocalCtx;
@@ -81,7 +80,7 @@ public record LittleTyper(@NotNull TyckState state, @NotNull LocalCtx localCtx) 
       case FormTerm.PartTy ty -> FormTerm.Univ.ZERO;
       case IntroTerm.PartEl el -> new FormTerm.PartTy(el.rhsType(), el.partial().restr());
       case FormTerm.Path path -> FormTerm.Univ.ZERO;
-      case IntroTerm.PathLam lam -> new FormTerm.Path(new Cube<>(
+      case IntroTerm.PathLam lam -> new FormTerm.Path(new FormTerm.Cube(
         lam.params().map(Term.Param::ref),
         term(lam.body()),
         new Partial.Const<>(term(lam.body()))
